refactor: rename i/e/n/mockablex => i/netxmocks
Needed to more easily do https://github.com/ooni/probe/issues/1505
