--- internal/engine/netx/mockablex/doc.go
@@ -1,2 +1,0 @@
-// Package mockable contains mocks for netx types.
-package mockablex
