+++ internal/netxmocks/doc.go
@@ -1,0 +1,2 @@
+// Package netxmocks contains mocks for netx types.
+package netxmocks
